app/ai/requirements.txt
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
langchain_ollama
langchain_chroma
wikipedia
langchain_groq
groq
beautifulsoup4